pyproject.toml
[build-system]
requires = ["hatchling>=1.24.0"]
build-backend = "hatchling.build"

[project]
name = "invoice-pdf-cli"
version = "0.1.0"
description = "CLI tool to extract invoice date and short description from PDFs using Gemini"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "google-genai>=1.0.0",
  "pydantic>=2.7.0",
  "pymupdf>=1.24.0",
  "typer>=0.12.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
]

[project.scripts]
invoice-extract = "invoice_extract_cli.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/invoice_extract_cli"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[tool.uv]
package = true
